Moved test_create_metadata_database_tables to common
