+++ crates/modelardb_common/Cargo.toml
@@ -29,3 +29,5 @@ sqlx = { workspace = true, features = ["runtime-tokio-rustls"] }
 
 [dev-dependencies]
 proptest.workspace = true
+tempfile.workspace = true
+tokio = { workspace = true, features = ["rt-multi-thread", "signal"] }
